creacion de la clase user, su repositorio y controlador para el manejo de usuarios ademas de verificaciones en el todo

+++ src/main/java/com/backend/todo/controller/ToDoController.java
@@ -26,7 +26,7 @@ public class ToDoController {
     }
 
     @GetMapping("/{taskId}")
-    public ResponseEntity<ToDo> getTaskById(@PathVariable Long taskId) {
+    public ResponseEntity<ToDo> getTaskById(@PathVariable int taskId) {
         Optional<ToDo> task = toDoRepository.findByTaskId(taskId);
         if (task.isPresent()) {
             return ResponseEntity.ok(task.get());
